backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
